Take over pdf generation, separator images
For a while, I had pdf generation in the epub library, but that wasn't
making a lot of sense, so I pulled the pdf generation into the cahaba
command. Then, I wanted to have large separators with images in them, so
I setup the asset folder and pulled out the css file and renamed it to
volume.css. Then just add page numbers and headers to pdfs and
that's all folks.